Update README.md
Hachiko Monogatari is a 1987 Japanese film that tells the heartwarming and tragic story of **Hachikō**, an Akita dog, who became famous for his remarkable loyalty to his owner, even after the owner's death. This film is based on the true story of Hachikō, who waited for his deceased owner at the Shibuya Station every day for nearly 10 years.